feat(activity): 完善活动报名功能并新增审核流程
- 新增ActivityReviewAction和ActivityReviewStage常量类用于审核管理
- 在ActivityWithScheduleAddForm中添加学院报名ID列表字段
- 修改GradeEntity中的canEnrollGrade字段类型从Integer改为Long
- 重构ActivityWithScheduleService中的活动添加逻辑，增加参数验证
- 实现新的addNotReviewedOne方法，支持活动审核流程
- 引入DangerousUserException枚举类型区分系统端口
- 在PortalUserManager中添加获取未删除用户的便捷方法
- 更新活动实体类引入List导入修复编译问题
- 集成敏感词过滤依赖库用于内容安全检查

+++ sa-admin/src/main/java/net/lab1024/sa/admin/module/business/funcampus/activityCanEnrollGrade/domain/entity/ActivityCanEnrollGradeEntity.java
@@ -5,6 +5,8 @@ import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import java.time.LocalDateTime;
+import java.util.List;
+
 import lombok.Data;
 
 /**
@@ -33,7 +35,7 @@ public class ActivityCanEnrollGradeEntity {
     /**
      * 能报名的年级
      */
-    private Integer canEnrollGrade;
+    private Long canEnrollGrade;
 
     /**
      * 是否删除
